Computer_Store/src/com/marc/store/service/imp/GoodsServiceImp.java
package com.marc.store.service.imp;

import com.marc.store.dao.GoodsDao;
import com.marc.store.dao.imp.GoodsDaoImpJdbc;
import com.marc.store.domain.Goods;
import com.marc.store.service.GoodsService;

import java.util.List;

public class GoodsServiceImp implements GoodsService {

    GoodsDao goodsDao = new GoodsDaoImpJdbc();

    @Override
    public List<Goods> queryAll() {
        return goodsDao.findAll();
    }

    @Override
    public List<Goods> queryByStartEnd(int start, int end) {
        return goodsDao.findStartEnd(start, end);
    }

    @Override
    public Goods querDetail(long goodsid) {
        return goodsDao.findByPk(goodsid);
    }
}
